requirements.txt
# CutoutAI Dependencies

# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
timm>=0.9.0
kornia>=0.7.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0

# Optional: for HuggingFace model loading
huggingface-hub>=0.19.0
